Extract node_runtime mock to a module.
