+++ ailets-rs/messages_to_markdown/tests/actor.rs
@@ -1,69 +1,7 @@
-use lazy_static::lazy_static;
 use messages_to_markdown::messages_to_markdown;
-use std::sync::Mutex;
 
-struct MockReadFile {
-    buffer: Vec<u8>,
-    pos: usize,
-}
-
-lazy_static! {
-    static ref MOCK_WRITE_FILE: Mutex<Vec<u8>> = Mutex::new(Vec::new());
-    static ref MOCK_READ_FILE: Mutex<MockReadFile> = Mutex::new(MockReadFile {
-        buffer: Vec::new(),
-        pos: 0,
-    });
-}
-
-fn clear_mocks() {
-    let mut file = MOCK_WRITE_FILE.lock().unwrap();
-    file.clear();
-}
-
-pub fn set_input(inputs: &[&str]) {
-    let mut file = MOCK_READ_FILE.lock().unwrap();
-    file.buffer.clear();
-    for input in inputs {
-        file.buffer.extend_from_slice(input.as_bytes());
-    }
-    file.pos = 0;
-}
-
-#[no_mangle]
-pub extern "C" fn n_of_streams(_name_ptr: *const u8) -> u32 {
-    1
-}
-
-#[no_mangle]
-pub extern "C" fn open_read(_name_ptr: *const u8, _index: u32) -> u32 {
-    0
-}
-
-#[no_mangle]
-pub extern "C" fn open_write(_name_ptr: *const u8) -> u32 {
-    0
-}
-
-#[no_mangle]
-pub extern "C" fn aread(_fd: u32, buffer_ptr: *mut u8, count: u32) -> u32 {
-    let mut file = MOCK_READ_FILE.lock().unwrap();
-    let buffer: &mut [u8] = unsafe { std::slice::from_raw_parts_mut(buffer_ptr, count as usize) };
-    let bytes_to_copy = std::cmp::min(count as usize, file.buffer.len() - file.pos);
-    buffer[..bytes_to_copy].copy_from_slice(&file.buffer[file.pos..file.pos + bytes_to_copy]);
-    file.pos += bytes_to_copy;
-    bytes_to_copy as u32
-}
-
-#[no_mangle]
-pub extern "C" fn awrite(_fd: u32, buffer_ptr: *mut u8, count: u32) -> u32 {
-    let mut file = MOCK_WRITE_FILE.lock().unwrap();
-    let buffer = unsafe { std::slice::from_raw_parts(buffer_ptr, count as usize) };
-    file.extend_from_slice(buffer);
-    count as u32
-}
-
-#[no_mangle]
-pub extern "C" fn aclose(_fd: u32) {}
+mod mocked_node_runtime;
+use mocked_node_runtime::{clear_mocks, get_output, set_input};
 
 #[test]
 fn test_basic_conversion() {
@@ -79,8 +17,7 @@ fn test_basic_conversion() {
 
     messages_to_markdown();
 
-    let file = MOCK_WRITE_FILE.lock().unwrap();
-    assert_eq!(&*file, b"Hello!");
+    assert_eq!(get_output(), "Hello!");
 }
 
 #[test]
@@ -99,8 +36,7 @@ fn test_multiple_content_items() {
 
     messages_to_markdown();
 
-    let file = MOCK_WRITE_FILE.lock().unwrap();
-    assert_eq!(&*file, b"First item\n\nSecond item\n\nThird item");
+    assert_eq!(get_output(), "First item\n\nSecond item\n\nThird item");
 }
 
 #[test]
@@ -124,8 +60,10 @@ fn test_two_messages() {
 
     messages_to_markdown();
 
-    let file = MOCK_WRITE_FILE.lock().unwrap();
-    assert_eq!(&*file, b"First message\n\nSecond message\n\nExtra text");
+    assert_eq!(
+        get_output(),
+        "First message\n\nSecond message\n\nExtra text"
+    );
 }
 
 #[test]
@@ -136,8 +74,7 @@ fn test_empty_input() {
 
     messages_to_markdown();
 
-    let file = MOCK_WRITE_FILE.lock().unwrap();
-    assert_eq!(&*file, b"");
+    assert_eq!(get_output(), "");
 }
 
 #[test]
@@ -159,6 +96,5 @@ fn test_long_text() {
 
     messages_to_markdown();
 
-    let file = MOCK_WRITE_FILE.lock().unwrap();
-    assert_eq!(&*file, long_text.as_bytes());
+    assert_eq!(get_output(), long_text);
 }
